Some changes
I've made the modules list command a single line, and a number showing how many valid modules there are. For example:
Modules (4): hello,help,say,cookies

I've also changed the invalid module warning to be more user-friendly.
Instead of "No valid command", it is "[module] is not a valid module! Use 'execute modules' for a list of modules."

I've also added a console log to show if a function hasn't been loaded because the selection of typeof() had deemed it unworthy of auto-loading. I suspect the function is wrongdoing because the auto-load doesn't get evaluated.